feat: added quantity and user can input the date of the habit

+++ HabitTracker.Mo3ses/Program.cs
@@ -1,4 +1,5 @@
 ﻿using HabitTracker.Mo3ses;
+using static System.Data.Entity.Infrastructure.Design.Executor;
 
 public class Program
 {
@@ -10,6 +11,7 @@ public class Program
         do
         {
             string inputValue;
+            bool habitChecker = false;
             Menu.ShowMenu();
             Console.Write("OPTION: ");
             inputValue = Console.ReadLine();
@@ -25,7 +27,9 @@ public class Program
                         Console.Write("CHOOSE HABIT ID YOU WANT TO RECORD: ");
                         inputValue = Console.ReadLine();
                         int trackId = ProgramHelpers.ValidateInputs(inputValue);
-                        conn.TrackHabit(trackId);
+
+                        habitChecker = conn.HasHabit(trackId);
+                        if (habitChecker) conn.TrackHabit(trackId);
                     }
                     break;
                 case "2":
@@ -54,7 +58,8 @@ public class Program
                         Console.Write("CHOOSE THE HABIT ID TO CHANGE: ");
                         inputValue = Console.ReadLine();
                         int updateId = ProgramHelpers.ValidateInputs(inputValue);
-                        if (updateId != 0)
+                        habitChecker = conn.HasHabit(updateId);
+                        if (updateId != 0 && habitChecker == true)
                         {
                             Console.Write("CHOOSE THE NEW HABIT NAME: ");
                             string newHabitName = Console.ReadLine();
@@ -62,7 +67,7 @@ public class Program
                             string newHabitMeasure = Console.ReadLine();
                             Console.Write("CHOOSE THE  NEW VALUE PER TRACK:");
                             string newHabitValue = Console.ReadLine();
-                            if (!string.IsNullOrEmpty(newHabitName) && !string.IsNullOrEmpty(newHabitMeasure) &&!(ProgramHelpers.ValidateInputs(newHabitValue) == 0))
+                            if (!string.IsNullOrEmpty(newHabitName) && !string.IsNullOrEmpty(newHabitMeasure) && !(ProgramHelpers.ValidateInputs(newHabitValue) == 0))
                             {
                                 conn.UpdateHabit(updateId, newHabitName, newHabitMeasure, ProgramHelpers.ValidateInputs(newHabitValue));
                             }
@@ -82,23 +87,66 @@ public class Program
                         Console.Write("CHOOSE THE HABIT ID YOU WANT TO DELETE: ");
                         inputValue = Console.ReadLine();
                         int deleteId = ProgramHelpers.ValidateInputs(inputValue);
-                        if (deleteId != 0)
+                        habitChecker = conn.HasHabit(deleteId);
+                        if (deleteId != 0 && habitChecker == true)
                         {
                             conn.DeleteHabit(deleteId);
                         }
                     }
                     break;
+
+
                     case "6":
+                    
+                    Menu.ShowReportMenu();
+                    inputValue = Console.ReadLine();
+                    int habitReportMenu = ProgramHelpers.ValidateInputs(inputValue);
+                   
                     bool hasHabitReport = conn.GetHabits();
+                    
                     if (hasHabitReport)
                     {
-                        Console.Write("CHOSE THE HABIT ID YOU WANT TO VIEW THE REPORT: ");
-                        inputValue = Console.ReadLine();
-                        int habitReportId = ProgramHelpers.ValidateInputs(inputValue);
-                        if (habitReportId != 0)
+                        
+
+                        switch (habitReportMenu)
                         {
-                            conn.HabitReport(habitReportId);
+                            case 0:
+                                break;
+                            case 1:
+                                Console.Write("CHOOSE THE HABIT ID YOU WANT TO VIEW THE REPORT: ");
+
+                                inputValue = Console.ReadLine();
+                                int habitReportId = ProgramHelpers.ValidateInputs(inputValue);
+                                habitChecker = conn.HasHabit(habitReportId);
+                                if (habitReportId != 0 && habitChecker == true)
+                                {
+                                    conn.HabitReport(habitReportId);
+                                }
+                                break;
+                             case 2:
+                                Console.Write("CHOOSE THE HABIT ID YOU WANT TO VIEW THE REPORT: ");
+                                inputValue = Console.ReadLine();
+                                int habitIntervalReportId = ProgramHelpers.ValidateInputs(inputValue);
+                                habitChecker = conn.HasHabit(habitIntervalReportId);
+                               
+                                if (habitIntervalReportId != 0 && habitChecker == true)
+                                {
+                                    Console.Write("CHOOSE START DATE (YYYY-MM-DD): ");
+                                    string habitStartDate = Console.ReadLine();
+
+
+                                    Console.Write("CHOOSE THE END DATE (YYYY-MM-DD): ");
+                                    string habitEndDate = Console.ReadLine();
+                                     
+
+                                    conn.HabitReportInterval(habitIntervalReportId, habitStartDate, habitEndDate);
+                                }
+                                break;
+                            default:
+                                break;
                         }
+
+                       
                     }
                     break;
                 default:
